image/requirements.txt
fake-useragent==1.4.0
Pillow==9.5.0
requests==2.31.0
